[TAAS-23] gss2json script now takes `--push` to send to github.
Rather than having a bunch of independent scripts, this gets our
`gss2json` code to have the extra smarts. With the `--push` switch it
now sends the data off to github.

This subsumes and improves upon our old `update.py` code. The Makefile
has been updated accordingly.
